refactor: moved exeption handling for callback_answer in EventInfo refactor: removed as much exception handling as possible
