+++ modules/debug/__init__.py
@@ -1,2 +1,2 @@
 """Modules used during debug"""
-from .log_manager import logger
+from .log_manager import logger, error_handler, log_message
